Moves miscellaneous types to the models package (#192)
