+++ models/api/models.api
@@ -1,10 +1,3 @@
-public final class com/revenuecat/purchases/kmp/DangerousSettings {
-	public fun <init> ()V
-	public fun <init> (Z)V
-	public synthetic fun <init> (ZILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public final fun getAutoSyncPurchases ()Z
-}
-
 public abstract interface class com/revenuecat/purchases/kmp/LogHandler {
 	public abstract fun d (Ljava/lang/String;Ljava/lang/String;)V
 	public abstract fun e (Ljava/lang/String;Ljava/lang/String;Ljava/lang/Throwable;)V
@@ -24,28 +17,9 @@ public final class com/revenuecat/purchases/kmp/LogLevel : java/lang/Enum {
 	public static fun values ()[Lcom/revenuecat/purchases/kmp/LogLevel;
 }
 
-public abstract interface class com/revenuecat/purchases/kmp/PurchasesAreCompletedBy {
-}
-
-public final class com/revenuecat/purchases/kmp/PurchasesAreCompletedBy$MyApp : com/revenuecat/purchases/kmp/PurchasesAreCompletedBy {
-	public fun <init> (Lcom/revenuecat/purchases/kmp/models/StoreKitVersion;)V
-	public final fun getStoreKitVersion ()Lcom/revenuecat/purchases/kmp/models/StoreKitVersion;
-}
-
-public final class com/revenuecat/purchases/kmp/PurchasesAreCompletedBy$RevenueCat : com/revenuecat/purchases/kmp/PurchasesAreCompletedBy {
-	public static final field INSTANCE Lcom/revenuecat/purchases/kmp/PurchasesAreCompletedBy$RevenueCat;
-	public fun equals (Ljava/lang/Object;)Z
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
-}
-
 public abstract interface class com/revenuecat/purchases/kmp/PurchasesDelegate {
 	public abstract fun onCustomerInfoUpdated (Lcom/revenuecat/purchases/kmp/models/CustomerInfo;)V
 	public abstract fun onPurchasePromoProduct (Lcom/revenuecat/purchases/kmp/models/StoreProduct;Lkotlin/jvm/functions/Function2;)V
-}
-
-public abstract interface class com/revenuecat/purchases/kmp/ReplacementMode {
-	public abstract fun getName ()Ljava/lang/String;
 }
 
 public final class com/revenuecat/purchases/kmp/UpdatedCustomerInfoDelegate : com/revenuecat/purchases/kmp/PurchasesDelegate {
@@ -93,6 +67,13 @@ public final class com/revenuecat/purchases/kmp/models/CustomerInfo {
 	public final fun getOriginalApplicationVersion ()Ljava/lang/String;
 	public final fun getOriginalPurchaseDateMillis ()Ljava/lang/Long;
 	public final fun getRequestDateMillis ()J
+}
+
+public final class com/revenuecat/purchases/kmp/models/DangerousSettings {
+	public fun <init> ()V
+	public fun <init> (Z)V
+	public synthetic fun <init> (ZILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun getAutoSyncPurchases ()Z
 }
 
 public final class com/revenuecat/purchases/kmp/models/DiscountPaymentMode : java/lang/Enum {
@@ -147,7 +128,7 @@ public final class com/revenuecat/purchases/kmp/models/EntitlementVerificationMo
 	public static fun values ()[Lcom/revenuecat/purchases/kmp/models/EntitlementVerificationMode;
 }
 
-public final class com/revenuecat/purchases/kmp/models/GoogleReplacementMode : java/lang/Enum, com/revenuecat/purchases/kmp/ReplacementMode {
+public final class com/revenuecat/purchases/kmp/models/GoogleReplacementMode : java/lang/Enum, com/revenuecat/purchases/kmp/models/ReplacementMode {
 	public static final field CHARGE_FULL_PRICE Lcom/revenuecat/purchases/kmp/models/GoogleReplacementMode;
 	public static final field CHARGE_PRORATED_PRICE Lcom/revenuecat/purchases/kmp/models/GoogleReplacementMode;
 	public static final field DEFERRED Lcom/revenuecat/purchases/kmp/models/GoogleReplacementMode;
@@ -313,6 +294,21 @@ public abstract interface class com/revenuecat/purchases/kmp/models/PromotionalO
 	public abstract fun getDiscount ()Lcom/revenuecat/purchases/kmp/models/StoreProductDiscount;
 }
 
+public abstract interface class com/revenuecat/purchases/kmp/models/PurchasesAreCompletedBy {
+}
+
+public final class com/revenuecat/purchases/kmp/models/PurchasesAreCompletedBy$MyApp : com/revenuecat/purchases/kmp/models/PurchasesAreCompletedBy {
+	public fun <init> (Lcom/revenuecat/purchases/kmp/models/StoreKitVersion;)V
+	public final fun getStoreKitVersion ()Lcom/revenuecat/purchases/kmp/models/StoreKitVersion;
+}
+
+public final class com/revenuecat/purchases/kmp/models/PurchasesAreCompletedBy$RevenueCat : com/revenuecat/purchases/kmp/models/PurchasesAreCompletedBy {
+	public static final field INSTANCE Lcom/revenuecat/purchases/kmp/models/PurchasesAreCompletedBy$RevenueCat;
+	public fun equals (Ljava/lang/Object;)Z
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
+}
+
 public final class com/revenuecat/purchases/kmp/models/PurchasesError {
 	public fun <init> (Lcom/revenuecat/purchases/kmp/models/PurchasesErrorCode;Ljava/lang/String;)V
 	public synthetic fun <init> (Lcom/revenuecat/purchases/kmp/models/PurchasesErrorCode;Ljava/lang/String;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
@@ -393,6 +389,10 @@ public final class com/revenuecat/purchases/kmp/models/RecurrenceMode : java/lan
 	public static fun getEntries ()Lkotlin/enums/EnumEntries;
 	public static fun valueOf (Ljava/lang/String;)Lcom/revenuecat/purchases/kmp/models/RecurrenceMode;
 	public static fun values ()[Lcom/revenuecat/purchases/kmp/models/RecurrenceMode;
+}
+
+public abstract interface class com/revenuecat/purchases/kmp/models/ReplacementMode {
+	public abstract fun getName ()Ljava/lang/String;
 }
 
 public final class com/revenuecat/purchases/kmp/models/Store : java/lang/Enum {
